Create algos module and move relevant files there
