+++ parley/src/algos/mod.rs
@@ -1,1 +1,9 @@
+// Copyright 2024 the Parley Authors
+// SPDX-License-Identifier: Apache-2.0 OR MIT
 
+// TODO - Remove pub(crate)
+
+pub(crate) mod bidi;
+pub(crate) mod resolve;
+pub(crate) mod shape;
+pub(crate) mod swash_convert;
