+++ parley/src/algos/shape.rs
@@ -1,9 +1,10 @@
 // Copyright 2021 the Parley Authors
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
+use crate::algos::resolve::{RangedStyle, ResolveContext, Resolved};
+use crate::algos::swash_convert::{locale_to_fontique, script_to_fontique, synthesis_to_swash};
 use crate::inputs::style::{Brush, FontFeature, FontVariation};
 use crate::layout::Layout;
-use crate::resolve::{RangedStyle, ResolveContext, Resolved};
 use crate::util::nearly_eq;
 use crate::Font;
 use fontique::QueryFamily;
@@ -248,8 +249,8 @@ impl<'a, 'b, B: Brush> FontSelector<'a, 'b, B> {
         let variations = rcx.variations(style.font_variations).unwrap_or(&[]);
         let features = rcx.features(style.font_features).unwrap_or(&[]);
         query.set_families(fonts.iter().copied());
-        let fb_script = crate::swash_convert::script_to_fontique(script);
-        let fb_language = locale.and_then(crate::swash_convert::locale_to_fontique);
+        let fb_script = script_to_fontique(script);
+        let fb_language = locale.and_then(locale_to_fontique);
         query.set_fallbacks(fontique::FallbackKey::new(fb_script, fb_language.as_ref()));
         query.set_attributes(attrs);
         Self {
@@ -306,7 +307,6 @@ impl<B: Brush> partition::Selector for FontSelector<'_, '_, B> {
         let mut selected_font = None;
         self.query.matches_with(|font| {
             if let Ok(font_ref) = skrifa::FontRef::from_index(font.blob.as_ref(), font.index) {
-                use crate::swash_convert::synthesis_to_swash;
                 use skrifa::MetadataProvider;
                 use swash::text::cluster::Status as MapStatus;
                 let charmap = font_ref.charmap();
