+++ parley/src/inputs/builder.rs
@@ -3,9 +3,9 @@
 
 //! Context for layout.
 
+use crate::algos::shape::shape_text;
 use crate::inputs::context::LayoutContext;
 use crate::inputs::style::{Brush, StyleProperty, TextStyle, WhiteSpaceCollapse};
-use crate::shape::shape_text;
 use crate::FontContext;
 
 use crate::layout::Layout;
